[optimize]build controllers & 調整為MVC架構
