+++ routes/modules/user.js
@@ -18,13 +18,7 @@ router.post("/:id", userController.getOtherUserData);
 router.post("/own/:id", authorization, userController.getOwnUserData);
 
 /** 個人-更新使用者資料 */
-router.patch(
-  "/own/:id",
-  authorization,
-  [validateEmail, validateAccount],
-  uploadFile.single("avatarFile"),
-  userController.updateUserData
-);
+router.patch( "/own/:id", authorization, [validateEmail, validateAccount], uploadFile.single("avatarFile"), userController.updateUserData);
 
 /** 個人-修改(背景)深色模式 */
 router.patch("/own/theme/:id", authorization, userController.setDarkMode);
